core: cooperate with raco exe
Previously, executables created with `raco exe' of applications that
use this library would fail due to the call to `collection-path'.
This change fixes that problem by installing the built so into the
Racket lib dir and tracking a runtime path to it rather than relying
on path construction at runtime.